+++ lib/shared/core/constants/shared_preferences_keys.dart
@@ -1,8 +1,10 @@
 class SharedPreferencesKeys {
   static const String repositoryPath = "repositoryPath";
   static const String repositoryBookmark = "repositoryBookmark";
+  static const String recentRepositoryPaths = "recentRepositoryPaths";
   static const String diffModeDisplay = 'diff_mode_display';
   static const String repositorySidebarWidth = 'repository_sidebar_width';
   static const String themeMode = 'theme_mode';
-  static const String branchGroupsExpansionState = 'branch_groups_expansion_state';
+  static const String branchGroupsExpansionState =
+      'branch_groups_expansion_state';
 }